Set max neighbor table (NDP/ARP shared) size for kodiak3
Summary:
**Problem:**
In resource accountant, NDP and ARP table is protected by the following configed limit:
```
DEFINE_int32(
    max_ndp_entries,
    4000,
    "Maximum NDP entries supported by Resource Accountant");

DEFINE_int32(
    max_arp_entries,
    1280,
    "Maximum ARP entries supported by Resource Accountant");
```
For most ASIC, the value is fine as the value is less than the ASIC limit. But for Kodiak3 (G200 ASIC), it is not true as:
```
G200/Yuba limit:
ndp-> 521 , arp -> 521 and ndp+arp -> 521
```

**Solution**
We do not want to lower the configured limit to be the Min among all ASICs, so neighbor limit check in resource accountant will be changed to use the limit of Min(configured, ASIC limit). And ndp count + arp count should be < unified neighbor table size limit if defined.

This diff defines the unified ASIC neighbor table size(so far for Yuba only).

Differential Revision: D80824545

fbshipit-source-id: 29dffb7e3196c4f8256037d8205b82987d1c3c81