Fix AgentFabricSwitchSelfLoopTest for Ramon3 FE13 mode
Summary:
In FE2 mode, it takes about 100 seconds to receive callbacks for all ports, so it's passing. But in FE13 mode,
it's now taking about 130+ seconds to receive callbacks for all ports and hence
it was failing. Increasing the timeout to 180 seconds to make it pass in FE13 mode also.

Now, in FE13 mode it's taking about 14.5 mins for the tests to complete, so within
netcastle timeout of 15mins.

Reviewed By: nivinl

Differential Revision: D81466012

fbshipit-source-id: 3efef56e40c42853531084b40a1b162c8c1771db