Reduce CI frequency to reduce spend
Summary: The FBOSS git repo has actions jobs that are racking up costs and putting us over budget. These are essential, and we need actions support to be reenabled. This should reduce the costs significantly by running the jobs at a schedule that is less frequent but still makes sense.

Reviewed By: shiva-menta

Differential Revision: D81352693

fbshipit-source-id: 2d41dc66ca2fc420cf8ac45b34000ac220d44146