Properly handle mux i2c adapter creation
Summary:
I2C mux adapters always live on a PCI adapter (potentially after extra levels of muxes), so creating a mux adapter also involves creating other adapters.

Previously we did not consider this when doing `registerAdapterForCleanup`. That led to failures in cases where a parent Adapter was created, and then never cleaned up because it wasn't registered.

To solve this: return the list of created adapters and then register them all for cleanup

Reviewed By: lohithchittineni

Differential Revision: D81187270

fbshipit-source-id: 0af8494503267de7ba6f63a6800c1a4351a3b0ca