log counter on delayed device creation
Summary:
To debug S554061 we need to get some more information on when devices are delayed. Add a simple counter here when a device takes >1 second to create - it should be much faster than that.

We will also add some verbose logging to the BSP code which happens between device creation triggering and when it actually finishes so we can have more insight into where the slowdown is.

This counter will help us know which devices to look at.

Reviewed By: somasun

Differential Revision: D81360563

fbshipit-source-id: 467663f1f1dd55d1a428df94aa30b004e967f3b1