Exclude Tos/DSCP check for BCM port span update mirror test
Summary: BCM port span udpate mirror tests are failing (AgentIngressPortSpanMirroringTest/IPv[46].UpdateSpanPortMirror) because the verification with `verifyPortMirrorProgrammed` is checking if the DSCP/Tos value is set on the mirror. However, for local port mirroring, this might not work. In SAI, there is no similar check for the DSCP/Tos value. Even if added, the SAI tests fail further confirming that DSCP/Tos may not be supported for local SPAN port mirroring. Remove the check for local SPAN port mirroring.

Reviewed By: nivinl

Differential Revision: D81339733

fbshipit-source-id: 6b1918a5a4e01b6a99960adff0cc21074f07580a