Update port in agent mirror update tests
Summary: For AgentIngressPort(Span|Erspan)MirroringTest/IPv[46].Update(Span|Erspan)PortMirror tests, update the port in addition to the DSCP. For SPAN mirroring, updating the DSCP may be a noop/irrelevant. The DSCP/Tos value might not be persisted between cold boot and warm boot. To accurately confirm updates to the mirror, also update the egress port for mirrored packets.

Reviewed By: maxwindiff

Differential Revision: D81367975

fbshipit-source-id: 092547403748b026bf2d01540c12791353951463