Clarify how to use T0/T1/T2 tests
Summary: In our oss wiki, we have a page detailing the T0/T1/T2 tests and what they are, as well as which tests fall under each category. But this leads vendors to question when to run which  category, and if these three categories represent the set of all tests. This blurb adds clarification so that vendors know how to use these tests.

Reviewed By: joseph5wu

Differential Revision: D81934577

fbshipit-source-id: e6d6c84b5198ed86713036157af6df4db302fecf