Assert for correct consolidation penalties for merged, candidate merge grps
Summary:
During tests we update, add, remove lots of candidate and merge
groups. At the end of each state update ensure that the consolidation
info is correct. We do so be recomputing the consolidation info
based on the current group configuration (nhops, route references etc)
and assert that at the end of all the state update operations the
info is as expected.

Separately we already assert for correctness of consolidation
info computation in UTs. These tests then complete the circle

Reviewed By: msomasundaran

Differential Revision:
D81097074

Privacy Context Container: L1297311

fbshipit-source-id: fa648c247bcbe121e168a4987c82676cfb0040f9