Skip oob_eeprom validation in fw_util
Summary:
fw_util config validator will fail on janga/tahan for oob_eeprom. We
currently have the entry in the config, but it's incomplete and we haven't
validated yet to make sure the logic in the config work or whether more changes
are needed. oob eeprom is not part of provisioning logic, so let's exclude it
for now in config validator so that we can fix a current failure on minerva
blades HW.

Reviewed By: ezeob002

Differential Revision: D81826430

fbshipit-source-id: 1263cf810bca749b84f1f3584bfadfdb2820c6c3