Mount fboss in Docker container instead of copying (#538)
Summary:
Pull Request resolved: https://github.com/facebook/fboss/pull/538

When building the fboss image, we used to copy the contents of the fboss repo on the host machine into the container. The user would have to re-build the image each time they make changes on their host machine in order to test them. We should instead mount the repository for ease of use and so that changes can persist on the host machine.

We still need to copy the fboss repo during image creation so we can use scripts for fetching dependencies, but this will be deleted afterwards so that the user will be using the mounted repository when entering the container.

Reviewed By: joseph5wu

Differential Revision: D81816484