Add flag to enable SDK dump
Summary:
Adds a flag to enable SDK dump, similar to sai tests (D81166731)

Unlike the sai_test implementation, this flag applies even if the test is successful.

The reasoning for this is so that we can collect both coldboot and warmboot state to help debug warmboot test failures, even when the coldboot version of the test passes.

This flag will only be used when --enable_debug_dumps is passed to the netcastle CLI

Reviewed By: shiva-menta

Differential Revision: D81601271

fbshipit-source-id: 89611e1bacfb384231ac2e6e5ae71cba291dc7b0