Fix the expect queue rate in the check of AgentCoppQosTest.CpuQueueHighRateTraffic
Summary:
Once fixing MT-831: CPU queue-low rate limit does not work for kodiak3, noticed AgentCoppQosTest.CpuQueueHighRateTraffic is disabled globally.

We should enable it so it can help to catch the issue in MT-831 earlier.

This is a small fix to use the correct expected rate limit in the test verify. Once this landed, we can start to enable this test for ASICs

Reviewed By: vasantspatil

Differential Revision: D82158561

fbshipit-source-id: 7cc7f97ffa65cd4d73686ec87e06a46083cfc103