Fix agent crash on SSWs
Summary:
As titled.

This diff fixes the issue introduced in D79678854.
Note that D79678854 is reverted to mitigate the SEV.

This diff introduces the change in D79678854 back with correct fix.
Refer D79678854 why this change is needed.

Only change done in this diff is to store the shared pointer to
MultiSwitchInterfaceMap (line #921), so that it doesn't get destroyed as a
result of stateUpdate, while we are walking
through interafces in the map and we have a interafce pointer that is freed (a
dangling pointer)

Reviewed By: daiwei1983, msomasundaran, jasmeetbagga

Differential Revision: D82316821

fbshipit-source-id: 60e6873334c1285e008ab9d1b48c40c9334c109a