aclErspan tests: fix ACL matching to use source port only
Summary:
Fixing the rest of the acl ingress and egress tests. Tested on th4 and th5.

ERSPAN test had a very strict ACL that didn't match the sent packets

Since src port was missing, SPAN tests were looping the packets and the mirrorPort count was very high compared to traffic port count (sent packets).

SAI api fix for egress mirror similar to the changes made for ingress D81138972.

TTL doesn't actually limit the packet looping as we saw in the span test so I removed it.

Added some comments to describe what's happening in the test

Differential Revision: D82166687

fbshipit-source-id: 32d4a80131c839220afeae6365f651c9e9e0792a