Remove leds from darwin fan_service config
Summary:
I discovered that the fan LEDs for darwin are managed by the driver itself, so fan_service has never been successfully setting the LEDs at all, because it can't.

This makes things a lot simpler since darwin's led interface is whack and we can't do anything about it until it gets on platform manager. Once that happens we'll be able to move the led management to userspace.

See D82330124 which uncovers that led sysfs file writes were always failing.

Technical details: The module parameter `managed_leds` in `rook-cpld-driver` is what controls this.
```
[root@fboss314860616.snc1 /sys/class/leds]# cat /sys/module/rook_fan_cpld/parameters/managed_leds
Y
[root@fboss314860616.snc1 /sys/class/leds]# echo 1 > /run/devmap/sensors/FAN_CPLD/fan1_led
-bash: echo: write error: Operation not permitted
[root@fboss314860616.snc1 /sys/class/leds]# echo N > /sys/module/rook_fan_cpld/parameters/managed_leds
[root@fboss314860616.snc1 /sys/class/leds]# echo 1 > /run/devmap/sensors/FAN_CPLD/fan1_led
[root@fboss314860616.snc1 /sys/class/leds]#
```

Reviewed By: somasun

Differential Revision: D82331852

fbshipit-source-id: 7bdf60d6aa29db04858ff2eb25be02e9ab378d52