Get/set for resetQueueCreditBalance in state
Summary:
**Context:**
The initial credit in VoQ switch ports allows a lot of packets to egress after TX is
disabled on the port. This results in inaccuracies in some tests which need TX to be
disabled before specific sequence is done. This new SAI attribute provides an option to
reset the credit balance on a per port basis which will stop TX on a port after ~400+
256B packets.

**In this diff:**
At titled.

Reviewed By: shri-khare

Differential Revision: D82701072

fbshipit-source-id: 5c3efbe7a667bc9072d249afee89955eef843bfa