New Port Extension Attribute to reset queue credit balance
Summary:
**Context:**
The initial credit in VoQ switch ports allows a lot of packets to egress after TX is
disabled on the port. This results in inaccuracies in some tests which need TX to be
disabled before specific sequence is done. This new SAI attribute provides an option to
reset the credit balance on a per port basis which will stop TX on a port after ~400+
256B packets.

**In this diff:**
Introduce the new extension attribute.

Reviewed By: shri-khare

Differential Revision: D82644936

fbshipit-source-id: 0a4b057cca14dd7d7c1db067302e157b64ad8e25