Add extra flags to wedge_qsfp_util command
Summary:
On some platforms like janga, we support multiple platform mappings. For agent and qsfp_service, the correct platform mapping is picked on some flags that are enabled in their respective configs. However, wedge_qsfp_util doesn't have a config and hence doesn't pick the correct platform mapping.

In this diff, I forward the janga_test flag to wedge_qsfp_util if the binary is running with that flag.

Reviewed By: shiva-menta

Differential Revision: D82686959

fbshipit-source-id: 769ae9756a1546be0691ad38aa2c214e7900def2