Define equals method for Resolved/Unresolved NextHops
Summary:
This is towards replacing flat_map with unordered hash map for next hop group member storage in SaiNextHopGroupManager https://fburl.com/code/n46r199q

While testing ECMP compression, we noticed that the group deletion performance is not ideal - further investigation led to the element removal in ref map, which internally was using flat map.

To fix the slowness, use unordered hash map instead for nhg members. To support that, we need to define hash and equals method for the key fields.

Reviewed By: jasmeetbagga

Differential Revision: D82776891

fbshipit-source-id: ac62280ed8632440d70a3395a19cddf42aa0d7ae