Explicitly instantiate adapted cow storage type to reduce agent build time
Summary:
Fsdb adapted sub manager is in the critical path of building fboss_sw_agent. To improve the build time, we need to improve adapted sub manager build time. We have already explored all possibilities of improving the underlying the libraries. Other possibilities are
1. Remove adapted sub manager and use cowstate sub manager so the predefined instantions are used.
2. Define explicit instantiation for cowstorage with the new adapted types so it can be reused in building adapted sub mgr

This diff implements the second option.

Differential Revision: D82672925

fbshipit-source-id: 4c4d216048568fcccd2afddfa35a2ab326c9ef8c