(refresh cycle) Add Logic for Ensuring Transceivers Don't Program Before PHY Programming
Summary:
Now that port-level logic and transceiver-level logic are in separate state machines, these two different state machines can operate independently.

However, one dependency that still exists between the state machines, is that transceivers cannot be programmed until all PHY programming is done – this is because we need the portToPortInfo struct, discussed in D77309354, to be populated, as the transceiver needs to be aware of the portProfiles it needs to program the transceiver to handle.

Thus, we need to add a logic gate of some kind on transceiver state machines to ensure they can't proceed with tcvr programming until its associated port state machines have finished PHY programming.

Summary of Changes:
- `TransceiverManager`
  - Adds new set of `TransceiverID`, `tcvrsReadyForProgramming`, that represents transceivers ready for transceiver programming.
  - `PortManager` uses a helper function `markTransceiverReadyForProgramming` to update once all associated ports of a transceiver have finished PHY programming.
  - We check if a transceiver is ready for programming (based on `tcvrsReadyForProgramming`) before enqueueing a `PROGRAM_TRANSCEIVER` event – this all happens within `programTransceiver`.
  - Clears transceiver from `tcvrsReadyForProgramming` - comment describes why we need to do this.
- `PortManager`
  - Calls `markTransceiverReadyForProgramming` in:
    - `programInternalPhyPorts` / `programExternalPhyPorts` to mark when the transceiverIDs belonging to portIDs are done with PHY programming (depending on if in XPHY system or not).

Reviewed By: harshitgulati18

Differential Revision: D77309356

fbshipit-source-id: 474c8a52990ccbab9640b67656be936df1c1eec5