(refresh cycle) Implement PortManager updateTransceiverPortStatus() + PortStatusCache Functions
Summary:
Summary of Changes: `updateTransceiverPortStatus()` is the primary refresh cycle function used for detecting changes to port state and trigger any corresponding state machine events based on:
- If port is newly enabled / disabled.
- If port is newly up / down.

This is a port-level operation, so should be now managed through `PortManager`. The function uses the same core logic, but given that port state is now stored in Port State Machines, we had the following changes:
- Initializing / uninitializing port state machines based on enabled status.
- Updating cache of enabled port statuses.

We also introduced an "initialized ports" cache here – all ports that have their state initialized and are in use – this is helpful to have, as seen in the next diffs, because many functions will be a lot quicker to iterate over a list of initialized ports.

Reviewed By: harshitgulati18

Differential Revision: D77309360

fbshipit-source-id: 55185138009efa0f5e05db548ddf0aaaaea01c96