Add APIs to keep track of link counts and validations for fabric link mon
Summary:
**Context:**
Fabric Link Monitoring provides a way to send packets on specific fabric links to
monitor the health of the links. This needed specific programming on J3 and R3
devices. The sending of packets/cells will be done from the local CPU of J3 for
RDSW/EDSW-FDSW links and the remote R3 will loopback the packets/cells. For
FDSW-SDSW links, the FDSW CPU will send the packets/cells and the SDSW side
will loop it back. Successful reception of packets/cells on the sender shows the link
is healthy. Detailed design of this feature captured [here](https://docs.google.com/presentation/d/1dKanpOhiMIhuPg-kD-RNtwUl3IDrwmubaGL_MgBTvVQ/edit?usp=sharing). This stack introduces diffs
to support this feature.

**In this diff:**
- New APIs to keep track of links between layers and validation.

Reviewed By: shri-khare

Differential Revision: D82391604

fbshipit-source-id: 9e0616243a96f5f302a063ae70a24ad87a12130f