API to get virtual device ID based on local or remote switch ports
Summary:
**Context:**
Fabric Link Monitoring provides a way to send packets on specific fabric links to
monitor the health of the links. This needed specific programming on J3 and R3
devices. The sending of packets/cells will be done from the local CPU of J3 for
RDSW/EDSW-FDSW links and the remote R3 will loopback the packets/cells. For
FDSW-SDSW links, the FDSW CPU will send the packets/cells and the SDSW side
will loop it back. Successful reception of packets/cells on the sender shows the link
is healthy. Detailed design of this feature captured [here](https://docs.google.com/presentation/d/1dKanpOhiMIhuPg-kD-RNtwUl3IDrwmubaGL_MgBTvVQ/edit?usp=sharing). This stack introduces diffs
to support this feature.

**In this diff:**

Find the remote VD associated with a port, both for VoQ switch and for fabric/spine
devices.

Reviewed By: shri-khare

Differential Revision: D82391599

fbshipit-source-id: 69d48aa2d1c806fb12616bda9defa5be449fe355