Bug fix, don't assume that new merge group creation always happens with grp add
Summary:
In case a new grp has the same nhops as a preexisting merged
group, we just fold the new group into the merged group and compute
candidate merges. In this case we don't see a new group add. So don't
assume this as a trigger for candidate merge update

Reviewed By: msomasundaran

Differential Revision:
D82700991

Privacy Context Container: L1297311

fbshipit-source-id: 3c816ae858573cd4531647e8c60ca85c1c62e19e