Check for amber or yellow LEDs
Summary:
We want to simplify LED naming and since the OCP specification refers to amber instead of yellow, we want to go with Amber. Right now we're checking specifically for "yellow" in the PM hw tests but this changes so we now accept either yellow or amber.

Eventually we will remove this and just have amber

Reviewed By: somasun, lohithchittineni

Differential Revision: D83158166

fbshipit-source-id: 808df63ef5978a452b1a1be338b869dee438d227