Change Tajo platforms to use Tajo as the product name
Summary:
I'm about to make some code refactor to allow AgentExecutor to be open source so that OSS can also use its logic to setup the muti-switch agent services.
I noticed some of the AgentNetWhoAmI functions are not using Tajo as the platform product name.
This diff is just trying to correct the functions names so that we can make it aligned with how we OSS Tajo Platform/Asic

Reviewed By: phshaikh

Differential Revision: D83515899

fbshipit-source-id: 92fae4fb227525c5351206dc457bca894d9353fa