Add sFlow SampleRecordOwned struct
Summary:
The sFlow specification (https://sflow.org/sflow_version_5.txt) uses opaque data structures to represent member variables of a struct. Currently in FBOSS, a list of opaque data structures is represented by the length of the list and then a pointer to the beginning of the first structure.

These changes modernize the code by encapsulating the lists by using vectors that own the data. First, a new structure will be added suffixed with "Owned". Next, implementations will use the new "Owned" versions. Finally, the old versions will be removed and the "Owned" suffix will be dropped.

Opaque data structures are serialized with a uint32 for the format specification and a uint32 for the length of the data of the opaque data structure.

Differential Revision: D83079089

fbshipit-source-id: 09b7aa7b37dc48b6a3b6f19f12c69a5675d284ef