Fix route metadata for ars scale mode for route modify
Summary: In ars scale mode, we set route metadata from nhgroup if it is non zero. For non zero case, we set it to default. Since metadata is optional, setting to default does not result in hw write. Set the route metadata to 0 if route is pointing to nexthopgroup and not using alternate. Also cache the previous metadata and set it to 0 if route moves from alternate nexthopgroup to single nexthop

Differential Revision: D83633238

fbshipit-source-id: 8400af7175c32e70c579f8f6c67453c1a3c0aa7d