Remove SW overtemp configuration for Tahan
Summary:
[FBOSS/Platform] Remove SW overtemp configuration for Tahan
Removing this logic from Fan Service config in Tahan.
The reason being that, the reading is sometimes inaccurate, and go beyond the threshold,
even when the TH5 ASIC is actually not that hot. Since the inaccurate reading is already
causing problem in Minerva testbeds, we are removing this config, to disable the feature.

We are still covered by HW over protection logic, which is a bit relaxed.

Reviewed By: binhuang00

Differential Revision: D83688684

fbshipit-source-id: c39e35da4e10807469e236b25d659bec2556edd3