Add AGERA XPHY topology traversal and pin junction logic
Summary:
We are updating the logic for get_connection_pairs_for_profile.

Now, when xphy is involved, there is a connection between xphy_line and either the backplane or transceiver. There is also a corresponding connection between npu and xphy_system.

The connection between xphy_system and xphy_line is not explicitly defined in the static mappings. Therefore, when we encounter a next_connection_end as an XPHY LINE end, we replace it with the corresponding XPHY SYSTEM connection to ensure proper traversal towards the NPU.

Additionally, we have added a find_corresponding_xphy_system_lane() function to  get the corresponding sys lane for a given line lane, using same slot_id, chip_id, core_id, and lane_id.

Reviewed By: harshitgulati18

Differential Revision: D83586249

fbshipit-source-id: 22054aeddb299312ab963cc39af4842669b6aa88