{easy} Add port name to am idles h/w test
Summary:
I was trying to enable the test for j3 and found that the test fails.

The reason was that unlike tahan whose first ports are front panel, on j3 the first ports are rcy.

This attr is not supported on rcy ports
https://www.internalfb.com/code/fbsource/[1e373b74a3f4f4afd6134095968c1a7c3d8b8e62]/third-party/tp2/brcm-sai/12.2.0.0_odp/brcm-sai-12.2.0.0_odp/src/brcm_sai_port.c?lines=4673-4693

Add this log helped identify this problem. Fix in next diff.

Reviewed By: jasmeetbagga

Differential Revision: D83673258

fbshipit-source-id: 6f53b127a8634971249679b94784c71fe1bedd71