Viper platform mapping to enable uniform local offset across dsf cluster
Summary:
System port id 0, 1, 2 and 3 are for cpu ports.

When we try to make the local offset for all switches uniform [0-10] (today local offset and global offset are same and unique for each swtich in a dsf cluster), the current logical port-ids for local rcy ports of 2, 3, .. leads to these ports getting assigned ssytem port-id of 2, 3..

This collides with system port-ids of cpu ports.

2 options - make local offset 3 (implemented here)

re-number rcy and evt ports from 2, 3, 4, 5 and 6 to 4, 5, 6 and 8

7 is stomped by be2fe (global), so can't be mvoed.

Differential Revision:
D83723259

Privacy Context Container: L1297311

fbshipit-source-id: 9ee1d8ad14d9e6c5a64590c0a2d8602dd2cf5e90