Fix L3 stats
Summary:
When we add routes (IPv4 or IPv6) into ASICs, its not always straightforward to calculate hardware resource usage stats. What I mean is, let's assume when no routes are added, we have N entries available. After adding 1 route, we expect N-1 entries to be available. This is true for most of the hardware resources, but not  for resources managed by LPM (Longest Prefix Match) algorithms.

This is because, vendors typically use TCAM and SRAM(buckets) in conjunction to add routes in hardware to achieve LPM. So, in this case, calculating number of free entries is always a heuristic approach. That is, number of routes we can add to the hardware will depend on the type of prefixes and the order in which we add them.

So, instead of just adding 1 IPv6 prefixes, we are adding 10 IPv6 prefixes, to make ensure that stats get updated(decremented).

Reviewed By: phshaikh

Differential Revision: D82617054

fbshipit-source-id: 08f7f1256157438d3e13cd68804dfbdb194841c2