Add unit test for vlan interface
Summary:
## Context

**High-Level Intention of the Stack of Diffs:**

The stack of diffs aims to enhance the Fboss Agent with the ability to manage and utilize desired peer information for interfaces, specifically in the context of GPU connectivity. The overall purpose of these diffs is to enable the configuration and automatic handling of neighbor solicitation for GPU interfaces when they become operational. This includes introducing new API methods for setting and retrieving desired peer information, populating this information from configuration, and modifying the agent's behavior to send neighbor solicitations when a GPU interface is created and becomes operational. The goal is to streamline the process of establishing connectivity between GPUs and improve the overall efficiency and reliability of the network.

## This Diff

This diff adds a new unit test for the VLAN interface in FBOSS Agent.

#### Changes

Two files are modified in this diff:

`fbcode/fboss/agent/test/PortUpdateHandlerNDPTest.cpp`:
   - This is a new file that contains a unit test for the VLAN interface.
   - The test case is added to check the PortUpdateHandler for NDP (Neighbor Discovery Protocol) functionality.
   - The test creates a `SwitchConfig` using the `createPortUpdateHandlerNDPTestConfig` function, which sets up the test environment.
   - The test then exercises and verifies the PortUpdateHandler's behavior for NDP updates.
 -  This unit tests verifies ndp cache, neighbor Solicitation retransmission mechanism

Reviewed By: ashutoshgrewal

Differential Revision: D82395588

fbshipit-source-id: ad55f113e2eebc04d669288fcb9f47e95c6415b3