Ndp cold start: send neighbor solicitiation once link is oper up
Summary:
#### Ndp Cold Start: Send Neighbor Solicitation Once Link is Oper Up

This diff makes modifications to the PortUpdateHandler class to handle the case where a port becomes operational. The changes aim to send a neighbor solicitation when a port comes up if DesiredPeerAddressIPv6 for GPU is configured.

*   **Implemented `handlePortUp` function in `PortUpdateHandler.cpp`:** The implementation of `handlePortUp` checks for associated interfaces when a port becomes operational and sends a neighbor solicitation as needed.
*  Neighbor cache takes care of retransmission, and retry till 300 times, before giving up. If agent doesn't receive response after 300 retries, it mark entry in neighbor cache is EXPIRED.

TODO: Need to check if we need to handle case when agent doesn't receive response after 300 retries. Do we keep on retrying ? or as we only start sending once port is UP, we should be ok.

Eventually I will adding feature flag for these changes, but for this diff I want ondiff to cover this code also. I will land all changes together.

Reviewed By: ashutoshgrewal

Differential Revision:
D82395571

Privacy Context Container: L1297311

fbshipit-source-id: e44acff189fb27362ea9e9b8c02d1b6a8b7e835c