Allow for computing penalty when no nhops are present
Summary:
As titled. Penalty for such a case would be 0, having
no nhops means you can't lose any nhops. While in prod we should
never have such a use case, crashing the process in a standalone
helper if it gets called as such is not a wise choice

Reviewed By: msomasundaran

Differential Revision:
D83783712

Privacy Context Container: L1297311

fbshipit-source-id: 73f64fb04bcef9c4e084f1f3167912f5b952b094