Add Discovered -> Transceiver_Ready Transition in Port Manager Mode
Summary:
Support state transition from DISCOVERED to TRANSCEIVER_READY, only in Port Manager mode. This was previously blocked due to our template depth limit, but now that is resolved in D83014061.

Previously, we "advanced" the state automatically to XPHY_PORTS_PROGRAMMED whenever a transceiver was discovered in PortManager mode, so that TRANSCEIVER_READY can technically be called on a transceiver that was just discovered.

However, this makes logging complicated, so this diff supports the new transition only when in PortManager mode.

Reviewed By: harshitgulati18

Differential Revision: D83252185

fbshipit-source-id: dcf78c8d98a20b637ec0a14cb3bff01c2be8d174