Refactor PhyManager Creation out of Platform-Specific Wedge Managers
Summary:
As titled – phyManager creations just depends on platform mapping, so there's no need for it to be in a platform-specific file when we initialize platform mapping.

This refactor is needed to instantiate phyManager separate from TcvrMgr, and pass it into TcvrMgr or PortMgr depending on if we're in port manager mode.

Note the build speed regression for `wedge-platform` – this is expected given that we now are importing all *PhyManager.h into `WedgeManagerInit.cpp`.

Reviewed By: harshitgulati18

Differential Revision: D83252184

fbshipit-source-id: 87bca3f286ed20406e7b5c0149cca0041c7473bb