Add setOverrideAllAgentPortStatusForTesting Variant
Summary:
For PortManager UTs, I rewrote `setOverrideAgentPortStatusForTesting` to have a different signature than TransceiverManager - this helped write cleaner tests.

However, qsfp hw tests, when run in PortManager mode, need a simple way to enable all relevant ports in order for transceivers to get programmed:
- In TransceiverManager mode, port enabled / disabled status is only relevant once transceivers get programmed. Transceivers can reach TRANSCEIVER_PROGRAMMED without agent status intervention.
- In PortManager mode, port enabled / disabled status is needed for transceivers to get programmed. PHY programming needs to be completed for transceivers to program, but PHY programming is only triggered once its associated port state machine is initialized by agent status ([LucidChart link](https://fburl.com/ntzrhbxe)).

Because of this, we need an easy way to tell PortManager to consider all ports (in `overrideTcvrToPortAndProfileForTest_`) as enabled to unblock all transceiver programming.

Reviewed By: harshitgulati18

Differential Revision: D83120696

fbshipit-source-id: 0a9755d2f412aa5962c47e46e42c608d7e9420f4