Add a new API to get switch ID from port ID
Summary:
**Context:**
Fabric Link Monitoring provides a way to send packets on specific fabric links to
monitor the health of the links. This needed specific programming on J3 and R3
devices. The sending of packets/cells will be done from the local CPU of J3 for
RDSW/EDSW-FDSW links and the remote R3 will loopback the packets/cells. For
FDSW-SDSW links, the FDSW CPU will send the packets/cells and the SDSW side
will loop it back. Successful reception of packets/cells on the sender shows the link
is healthy. Detailed design of this feature captured [here](https://docs.google.com/presentation/d/1dKanpOhiMIhuPg-kD-RNtwUl3IDrwmubaGL_MgBTvVQ/edit?usp=sharing). This stack introduces diffs
to support this feature.

**In this diff:**
Need to set the system port object associated with fabric link monitoring system ports
on the fabric ports. For this, we need to be able to derive the PortID corresponding to a
system port ID and hence the new util API.

Reviewed By: shri-khare

Differential Revision: D82898636

fbshipit-source-id: 0b085ea2e4b65ff0bcf48d11503152cfe0015b89