Compute the switch ID offset based on local/remote switch IDs and level
Summary:
**Context:**
Fabric Link Monitoring provides a way to send packets on specific fabric links to
monitor the health of the links. This needed specific programming on J3 and R3
devices. The sending of packets/cells will be done from the local CPU of J3 for
RDSW/EDSW-FDSW links and the remote R3 will loopback the packets/cells. For
FDSW-SDSW links, the FDSW CPU will send the packets/cells and the SDSW side
will loop it back. Successful reception of packets/cells on the sender shows the link
is healthy. Detailed design of this feature captured [here](https://docs.google.com/presentation/d/1dKanpOhiMIhuPg-kD-RNtwUl3IDrwmubaGL_MgBTvVQ/edit?usp=sharing). This stack introduces diffs
to support this feature.

**In this diff:**
Define a mechanism to generate a switchId offset based on local and remote switch
IDs. The computation is done differently for rdsw-fdsw connections and fsdw-sdsw
connections.

Reviewed By: shri-khare

Differential Revision: D82494502

fbshipit-source-id: bb7b40195945991ca82baad1766d1ec0244e0ab8