Add API to figure out parallel links and sequencing of links for switch ID allocation
Summary:
**Context:**
Fabric Link Monitoring provides a way to send packets on specific fabric links to
monitor the health of the links. This needed specific programming on J3 and R3
devices. The sending of packets/cells will be done from the local CPU of J3 for
RDSW/EDSW-FDSW links and the remote R3 will loopback the packets/cells. For
FDSW-SDSW links, the FDSW CPU will send the packets/cells and the SDSW side
will loop it back. Successful reception of packets/cells on the sender shows the link
is healthy. Detailed design of this feature captured [here](https://docs.google.com/presentation/d/1dKanpOhiMIhuPg-kD-RNtwUl3IDrwmubaGL_MgBTvVQ/edit?usp=sharing). This stack introduces diffs
to support this feature.

**In this diff:**
Handle parallel links in a VD in the topology. In case of parallel links, create an ordering
of local ports such that the offset for a specific link is the same on both sides of the
link.

Reviewed By: shri-khare

Differential Revision: D82391603

fbshipit-source-id: 34d00049ebcbc06f9bc880b52274d0c82fd86644