Move fabric link mon object to ThriftConfigApplier for reuse
Summary:
**Context:**
Fabric Link Monitoring provides a way to send packets on specific fabric links to
monitor the health of the links. This needed specific programming on J3 and R3
devices. The sending of packets/cells will be done from the local CPU of J3 for
RDSW/EDSW-FDSW links and the remote R3 will loopback the packets/cells. For
FDSW-SDSW links, the FDSW CPU will send the packets/cells and the SDSW side
will loop it back. Successful reception of packets/cells on the sender shows the link
is healthy. Detailed design of this feature captured [here](https://docs.google.com/presentation/d/1dKanpOhiMIhuPg-kD-RNtwUl3IDrwmubaGL_MgBTvVQ/edit?usp=sharing). This stack introduces diffs
to support this feature.

**In this diff:**
Fabric link monitoring object is needed at multiple places for a new config to figure the
system port ID and switch ID etc. and hence create the object once and use it as
needed in the apply thrift context.

Reviewed By: zechengh09, shri-khare

Differential Revision: D82898642

fbshipit-source-id: a85c8b04a9cc6996047c33d5ce3fcc54da3e4506