Helper functions to get the parallel link offset and lowest local switch ID
Summary:
**Context:**
Fabric Link Monitoring provides a way to send packets on specific fabric links to
monitor the health of the links. This needed specific programming on J3 and R3
devices. The sending of packets/cells will be done from the local CPU of J3 for
RDSW/EDSW-FDSW links and the remote R3 will loopback the packets/cells. For
FDSW-SDSW links, the FDSW CPU will send the packets/cells and the SDSW side
will loop it back. Successful reception of packets/cells on the sender shows the link
is healthy. Detailed design of this feature captured [here](https://docs.google.com/presentation/d/1dKanpOhiMIhuPg-kD-RNtwUl3IDrwmubaGL_MgBTvVQ/edit?usp=sharing). This stack introduces diffs
to support this feature.

**In this diff:**
Helper API to give the parallel link offset from the pre-computed ordered list. This is
applicable only in case we have parallel links in the topology to the same VD, which is
not currently the case.

Reviewed By: shri-khare

Differential Revision: D82422166

fbshipit-source-id: e4239e850ea8a1c841a94223b2f15790eedc21b2