Add config for system port offset for use in fab link monitoring
Summary:
**Context:**
Fabric Link Monitoring provides a way to send packets on specific fabric links to
monitor the health of the links. This needed specific programming on J3 and R3
devices. The sending of packets/cells will be done from the local CPU of J3 for
RDSW/EDSW-FDSW links and the remote R3 will loopback the packets/cells. For
FDSW-SDSW links, the FDSW CPU will send the packets/cells and the SDSW side
will loop it back. Successful reception of packets/cells on the sender shows the link
is healthy. Detailed design of this feature captured [here](https://docs.google.com/presentation/d/1dKanpOhiMIhuPg-kD-RNtwUl3IDrwmubaGL_MgBTvVQ/edit?usp=sharing). This stack introduces diffs
to support this feature.

**In this diff:**

Fabric link monitoring needs 160 system ports created on VoQ switches. The range of
system ports to use varies by deployment and hence will be passed in as an offset in
the config. This offset can be added to the fabric port ID to derive the system port to
use.

For example, in KCM2 dc, the fabricLinkMonitoringSystemPortOffset=-1015, so that we
end up allocating -1015 + 1024 = 9 for the first fabric port and -1015 + 1183 = 168 for
the last fabric port.

Reviewed By: shri-khare

Differential Revision: D81844169

fbshipit-source-id: 05c62a605f8c2d79a353f1a28b1fab7f04a4e47d