Shel Manager to also track ECMP routes to remote intf
Summary:
In order to report the accurate information of ECMP nexthops pointing to SHEL down ports, we also need to maintain the ref count for remote interfaces.

Imagine on RDSW, ECMP nexthops will point to remote interfaces on EDSWs (remote). It needs to keep track of remote intf ref count to report if any ECMP nexthop is resolved to EDSW interfaces.

Reviewed By: ashutoshgrewal

Differential Revision: D80500751

fbshipit-source-id: 2d64ebe14d92353933990e3dc2fae29de8e4462a