comment out the call to shell in sai replayer
Summary:
During the replay, if a call to shell is logged, the original code triggers shell and pauses for user inputs, which can be inconvenient during debugging.

The goal of this diff is to log the shell call as a C comment, thereby avoiding this issue while preserving the information.

Reviewed By: daiwei1983

Differential Revision: D81541611

fbshipit-source-id: dc879fe9075456db42915824be7c26e8ab3ef02a