Add logging to monitor delays in publishing switch state updates to FSDB
Summary:
Currently, we dont have a way to measure the delay between agent processing the switch state update -> publishing the update to FSDB. Create a utility to monitor and print these.

This utility can be reused by any FSDB publishers. QSFP for ex is publishing at a high rate and so might not desire to print these. Hence add a flag (false by default) which clients can set if they wish to print the delay

Reviewed By: zechengh09

Differential Revision: D83272786

fbshipit-source-id: cc283834792e746e8b882204ca7272408556bf1e