fix fboss2 show route details output
Summary:
Currently, "fboss2 show route details" is broken on fx switches with below incomplete output:
 {P1991353454}

This is because fboss2 is using wrong *ifNamePtr (could be a string like "fboss2001") when access vlanAggregatePortMap, which is expecting vlanId "2001". This diff resolves this issue by passing in vlanId.

Differential Revision:
D84566469

Privacy Context Container: L1297311

fbshipit-source-id: 88ae1932fe3e0c26c9318b222cca87dcb3c49caf