getChannelNumFromFrequency utility addition
Summary: This diff adds a utility function to the CmisModule class in the qsfp_service module of the FBOSS project. The function takes in a frequency in MHz and a FrequencyGrid enum and returns the corresponding channel number. The function uses a tunable module constexpr to define the default frequency and a conversion factor to convert MHz to GHz.

Reviewed By: xanabani

Differential Revision: D84538334

fbshipit-source-id: 03bc1091ee3da94e6cc66f7f623a993ad377746a