Handle 400G->200G transition on 400G-FR4 optics
Summary:
When we go from 400G -> 200G, we are going from a 8 lane datapath to 4 lane datapath. Previously we were releasing lanes 5-8 (programming app sel 0) but with D83919301 we stopped doing that. Thus when we configure lanes 1-4 in 200G, lanes 5-8 remain in the previous 400G mode and that's an invalid combination

In this diff, i bring back the lanes release logic. We ideally don't need this logic if we start supporting a map of valid SMF combinations for 400G/200G optics (we only do it for OSFP optics today). But that change is more complicated than bringing back the release logic, so going with this a temporary (medium term) fix.

Reviewed By: shiva-menta

Differential Revision: D84668493

fbshipit-source-id: a2bcbf74151c6dd436e63fe0a85e91dbaf66af2e