Properly unpack stable hashes for pinned dependencies
When the build starts it typically uses pinned dependencies from 
`fboss/oss/stable_commits/latest_stable_hashes.tar.gz` however due to a 
bug introduced in #538 the stable hashes were not extracted and were 
actually deleted, leading the build to expressly NOT use stable hashes 
when it was in fact supposed to.

The consequence of this is that builds were typically relying on the 
latest upstream sha1 for the various dependencies fetched from GitHub, 
which lead to unstable and non reproducible builds.